Partially organize source code
