+++ app/src/main/java/com/example/carcompanion/ui/car_info/car_details/CarDetailViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.carcompanion.ui.car_info
+package com.example.carcompanion.ui.car_info.car_details
 
 import android.annotation.SuppressLint
 import android.view.View
